Remove parser and rimport tests.
They are covered by integration tests.
