--- internal/engine/rimport/testdata/remote.json
@@ -1,344 +1,0 @@
-[
-  {
-    "Action": {
-      "AddLabel": "",
-      "Category": "",
-      "Archive": false,
-      "Delete": true,
-      "MarkImportant": false,
-      "MarkNotImportant": false,
-      "MarkRead": false,
-      "MarkNotSpam": false,
-      "Star": false
-    },
-    "Criteria": {
-      "From": "",
-      "To": "myselfyeah0@gmail.com",
-      "Subject": "",
-      "Query": ""
-    }
-  },
-  {
-    "Action": {
-      "AddLabel": "",
-      "Category": "",
-      "Archive": false,
-      "Delete": true,
-      "MarkImportant": false,
-      "MarkNotImportant": false,
-      "MarkRead": false,
-      "MarkNotSpam": false,
-      "Star": false
-    },
-    "Criteria": {
-      "From": "",
-      "To": "",
-      "Subject": "{\"[ExpatRepublicAmsterdam] 9 EURO YOGA\" \"(Amsterdam Football and Basketball)\"}",
-      "Query": ""
-    }
-  },
-  {
-    "Action": {
-      "AddLabel": "NotInteresting",
-      "Category": "",
-      "Archive": true,
-      "Delete": false,
-      "MarkImportant": false,
-      "MarkNotImportant": true,
-      "MarkRead": true,
-      "MarkNotSpam": false,
-      "Star": false
-    },
-    "Criteria": {
-      "From": "",
-      "To": "{myselfyeah0@gmail.com myselfyeah@gmail.com}",
-      "Subject": "",
-      "Query": ""
-    }
-  },
-  {
-    "Action": {
-      "AddLabel": "",
-      "Category": "",
-      "Archive": false,
-      "Delete": true,
-      "MarkImportant": false,
-      "MarkNotImportant": false,
-      "MarkRead": false,
-      "MarkNotSpam": false,
-      "Star": false
-    },
-    "Criteria": {
-      "From": "{*@onstageweb.com noreply@vivigratis.com}",
-      "To": "",
-      "Subject": "",
-      "Query": ""
-    }
-  },
-  {
-    "Action": {
-      "AddLabel": "",
-      "Category": "personal",
-      "Archive": false,
-      "Delete": false,
-      "MarkImportant": true,
-      "MarkNotImportant": false,
-      "MarkRead": false,
-      "MarkNotSpam": false,
-      "Star": false
-    },
-    "Criteria": {
-      "From": "foobar@gmail.com",
-      "To": "",
-      "Subject": "",
-      "Query": ""
-    }
-  },
-  {
-    "Action": {
-      "AddLabel": "BG",
-      "Category": "",
-      "Archive": false,
-      "Delete": false,
-      "MarkImportant": false,
-      "MarkNotImportant": false,
-      "MarkRead": false,
-      "MarkNotSpam": false,
-      "Star": false
-    },
-    "Criteria": {
-      "From": "{info@reservation.foo.co.uk *@bar.com}",
-      "To": "",
-      "Subject": "",
-      "Query": ""
-    }
-  },
-  {
-    "Action": {
-      "AddLabel": "Rust",
-      "Category": "",
-      "Archive": false,
-      "Delete": false,
-      "MarkImportant": false,
-      "MarkNotImportant": false,
-      "MarkRead": false,
-      "MarkNotSpam": false,
-      "Star": false
-    },
-    "Criteria": {
-      "From": "",
-      "To": "",
-      "Subject": "",
-      "Query": "list:zbaz.122677.list-id.mcsv.net"
-    }
-  },
-  {
-    "Action": {
-      "AddLabel": "Amici",
-      "Category": "personal",
-      "Archive": false,
-      "Delete": false,
-      "MarkImportant": false,
-      "MarkNotImportant": false,
-      "MarkRead": false,
-      "MarkNotSpam": false,
-      "Star": false
-    },
-    "Criteria": {
-      "From": "{zlskdj@gmail.com fsjei8.5@gmail.com jflgija@gmail.com}",
-      "To": "",
-      "Subject": "",
-      "Query": ""
-    }
-  },
-  {
-    "Action": {
-      "AddLabel": "DBLab",
-      "Category": "",
-      "Archive": false,
-      "Delete": false,
-      "MarkImportant": false,
-      "MarkNotImportant": false,
-      "MarkRead": false,
-      "MarkNotSpam": false,
-      "Star": false
-    },
-    "Criteria": {
-      "From": "*@dblab.it",
-      "To": "",
-      "Subject": "",
-      "Query": ""
-    }
-  },
-  {
-    "Action": {
-      "AddLabel": "",
-      "Category": "personal",
-      "Archive": false,
-      "Delete": false,
-      "MarkImportant": true,
-      "MarkNotImportant": false,
-      "MarkRead": false,
-      "MarkNotSpam": false,
-      "Star": false
-    },
-    "Criteria": {
-      "From": "express@airbnb.com",
-      "To": "",
-      "Subject": "",
-      "Query": "RE:"
-    }
-  },
-  {
-    "Action": {
-      "AddLabel": "Rust",
-      "Category": "",
-      "Archive": false,
-      "Delete": false,
-      "MarkImportant": false,
-      "MarkNotImportant": false,
-      "MarkRead": false,
-      "MarkNotSpam": false,
-      "Star": false
-    },
-    "Criteria": {
-      "From": "internals@rust-lang.org",
-      "To": "",
-      "Subject": "",
-      "Query": ""
-    }
-  },
-  {
-    "Action": {
-      "AddLabel": "Uni",
-      "Category": "",
-      "Archive": false,
-      "Delete": false,
-      "MarkImportant": false,
-      "MarkNotImportant": false,
-      "MarkRead": false,
-      "MarkNotSpam": false,
-      "Star": false
-    },
-    "Criteria": {
-      "From": "{*@univr.it flalsg.googlegroups.com}",
-      "To": "",
-      "Subject": "",
-      "Query": ""
-    }
-  },
-  {
-    "Action": {
-      "AddLabel": "Relocation",
-      "Category": "personal",
-      "Archive": false,
-      "Delete": false,
-      "MarkImportant": false,
-      "MarkNotImportant": false,
-      "MarkRead": false,
-      "MarkNotSpam": false,
-      "Star": false
-    },
-    "Criteria": {
-      "From": "{*@fsfl.com Folijfl@fslh.com iglfj@moving.com}",
-      "To": "",
-      "Subject": "",
-      "Query": ""
-    }
-  },
-  {
-    "Action": {
-      "AddLabel": "Relocation",
-      "Category": "personal",
-      "Archive": false,
-      "Delete": false,
-      "MarkImportant": false,
-      "MarkNotImportant": false,
-      "MarkRead": false,
-      "MarkNotSpam": false,
-      "Star": false
-    },
-    "Criteria": {
-      "From": "{Loo.afl@cartus.com bar.lkji@gmx.de}",
-      "To": "",
-      "Subject": "",
-      "Query": ""
-    }
-  },
-  {
-    "Action": {
-      "AddLabel": "",
-      "Category": "promotions",
-      "Archive": false,
-      "Delete": false,
-      "MarkImportant": false,
-      "MarkNotImportant": false,
-      "MarkRead": false,
-      "MarkNotSpam": false,
-      "Star": false
-    },
-    "Criteria": {
-      "From": "",
-      "To": "myself.yeah+ah@gmail.com",
-      "Subject": "",
-      "Query": ""
-    }
-  },
-  {
-    "Action": {
-      "AddLabel": "Uni",
-      "Category": "",
-      "Archive": false,
-      "Delete": false,
-      "MarkImportant": false,
-      "MarkNotImportant": false,
-      "MarkRead": false,
-      "MarkNotSpam": false,
-      "Star": false
-    },
-    "Criteria": {
-      "From": "",
-      "To": "asdflkj@studenti.uni.it",
-      "Subject": "",
-      "Query": ""
-    }
-  },
-  {
-    "Action": {
-      "AddLabel": "Bright",
-      "Category": "",
-      "Archive": false,
-      "Delete": false,
-      "MarkImportant": false,
-      "MarkNotImportant": false,
-      "MarkRead": false,
-      "MarkNotSpam": false,
-      "Star": false
-    },
-    "Criteria": {
-      "From": "*@brightfoobar.com",
-      "To": "",
-      "Subject": "",
-      "Query": ""
-    }
-  },
-  {
-    "Action": {
-      "AddLabel": "FOO",
-      "Category": "",
-      "Archive": false,
-      "Delete": false,
-      "MarkImportant": false,
-      "MarkNotImportant": false,
-      "MarkRead": false,
-      "MarkNotSpam": false,
-      "Star": false
-    },
-    "Criteria": {
-      "From": "{*@foo.it m.yeah@pec.it}",
-      "To": "",
-      "Subject": "",
-      "Query": ""
-    }
-  }
-]
